moved all classes to new package BJS2_68825
