+++ src/main/java/school/faang/BJS2_68825/HogwartsSpells.java
@@ -1,4 +1,4 @@
-package school.faang;
+package school.faang.BJS2_68825;
 
 import java.util.ArrayList;
 import java.util.HashMap;
