+++ src/main/java/school/faang/BJS2_68825/Main.java
@@ -1,4 +1,4 @@
-package school.faang;
+package school.faang.BJS2_68825;
 
 public class Main {
     public static void main(String[] args) {
